+++ graylog2-web-interface/src/components/indices/IndexSetFieldTypeProfiles/CreateProfileButton.tsx
