+++ graylog2-web-interface/src/components/indices/IndexSetFieldTypeProfiles/ProfileForm.test.tsx
@@ -22,16 +22,15 @@ import asMock from 'helpers/mocking/AsMock';
 import { loadViewsPlugin, unloadViewsPlugin } from 'views/test/testViewsPlugin';
 import useFieldTypesForMappings from 'views/logic/fieldactions/ChangeFieldType/hooks/useFieldTypesForMappings';
 import useFieldTypes from 'views/logic/fieldtypes/useFieldTypes';
-import EditProfile from 'components/indices/IndexSetFiledTypeProfiles/EditProfile';
-import useProfileMutations from 'components/indices/IndexSetFiledTypeProfiles/hooks/useProfileMutations';
+import ProfileForm from 'components/indices/IndexSetFieldTypeProfiles/ProfileForm';
 import { simpleFields } from 'fixtures/fields';
 import { profile1 } from 'fixtures/indexSetFieldTypeProfiles';
 
-const renderEditProfile = () => render(
-  <EditProfile profile={profile1} />,
+const mockSubmit = jest.fn();
+const mockCancel = jest.fn();
+const renderProfileForm = ({ initialValues }) => render(
+  <ProfileForm onCancel={mockCancel} onSubmit={mockSubmit} submitLoadingText="Submitting..." submitButtonText="Submit" initialValues={initialValues} />,
 );
-
-jest.mock('components/indices/IndexSetFiledTypeProfiles/hooks/useProfileMutations', () => jest.fn());
 jest.mock('views/logic/fieldactions/ChangeFieldType/hooks/useFieldTypesForMappings', () => jest.fn());
 
 jest.mock('views/logic/fieldtypes/useFieldTypes', () => jest.fn());
@@ -43,9 +42,6 @@ const selectItem = async (select: HTMLElement, option: string | RegExp) => {
 };
 
 describe('IndexSetFieldTypesList', () => {
-  const createMock = jest.fn(() => Promise.resolve());
-  const editMock = jest.fn(() => Promise.resolve());
-
   beforeAll(loadViewsPlugin);
 
   afterAll(unloadViewsPlugin);
@@ -58,57 +54,40 @@ describe('IndexSetFieldTypesList', () => {
           int: 'Number(int)',
           bool: 'Boolean',
           ip: 'IP',
+          date: 'Date',
         },
       },
       isLoading: false,
     });
-
-    asMock(useProfileMutations).mockReturnValue(({
-      editProfile: editMock,
-      isEditLoading: false,
-      createProfile: createMock,
-      isCreateLoading: false,
-      isLoading: false,
-    }));
 
     asMock(useFieldTypes).mockImplementation(() => (
       { data: simpleFields().toArray(), refetch: jest.fn() }
     ));
   });
 
-  it('Run editProfile with changed form data', async () => {
-    renderEditProfile();
-
-    const name = await screen.findByRole('textbox', {
-      name: /name/i,
-      hidden: true,
+  it('Do not run onSubmit when has empty name', async () => {
+    renderProfileForm({
+      initialValues: {
+        ...profile1,
+        name: '',
+      },
     });
 
-    const fieldFirst = await screen.findByLabelText(/select customFieldMappings.0.field/i);
-    const typeFirst = await screen.findByLabelText(/select customFieldMappings.0.type/i);
-    const submitButton = await screen.findByTitle(/update profile/i);
+    const submitButton = await screen.findByLabelText('Submit');
+    fireEvent.click(submitButton);
 
-    // eslint-disable-next-line testing-library/no-unnecessary-act
-    await act(async () => {
-      fireEvent.change(name, { target: { value: 'Profile 1 new name' } });
-      await selectItem(fieldFirst, 'date');
-      await selectItem(typeFirst, 'String type');
-      fireEvent.click(submitButton);
-    });
-
-    expect(editMock).toHaveBeenCalledWith({
-      name: 'Profile 1 new name',
-      description: 'Description 1',
-      id: '111',
-      customFieldMappings: [
-        { field: 'date', type: 'string' },
-        { field: 'user_ip', type: 'ip' },
-      ],
-    });
+    expect(mockSubmit).not.toHaveBeenCalled();
   });
 
-  it('Run editProfile with added form data', async () => {
-    renderEditProfile();
+  it('Do not run onSubmit when has empty customFieldMapping', async () => {
+    renderProfileForm({
+      initialValues: {
+        ...profile1,
+        customFieldMappings: [
+          { field: 'http_method', type: 'string' },
+        ],
+      },
+    });
 
     const addMappingButton = await screen.findByRole('button', { name: /add mapping/i });
 
@@ -117,26 +96,48 @@ describe('IndexSetFieldTypesList', () => {
       fireEvent.click(addMappingButton);
     });
 
-    const fieldThird = await screen.findByLabelText(/select customFieldMappings.2.field/i);
-    const typeThird = await screen.findByLabelText(/select customFieldMappings.2.type/i);
-    const submitButton = await screen.findByTitle(/update profile/i);
+    const typeSecond = await screen.findByLabelText(/select customFieldMappings.1.type/i);
+    const submitButton = await screen.findByLabelText('Submit');
 
     // eslint-disable-next-line testing-library/no-unnecessary-act
     await act(async () => {
-      await selectItem(fieldThird, 'date');
-      await selectItem(typeThird, 'String type');
+      await selectItem(typeSecond, 'String type');
+
       fireEvent.click(submitButton);
     });
 
-    expect(editMock).toHaveBeenCalledWith({
-      name: 'Profile 1',
-      description: 'Description 1',
-      id: '111',
-      customFieldMappings: [
-        { field: 'http_method', type: 'string' },
-        { field: 'user_ip', type: 'ip' },
-        { field: 'date', type: 'string' },
-      ],
+    expect(mockSubmit).not.toHaveBeenCalled();
+  });
+
+  it('Do not run onSubmit when has same fields in customFieldMapping', async () => {
+    renderProfileForm({
+      initialValues: {
+        ...profile1,
+        customFieldMappings: [
+          { field: 'http_method', type: 'string' },
+        ],
+      },
     });
+
+    const addMappingButton = await screen.findByRole('button', { name: /add mapping/i });
+
+    // eslint-disable-next-line testing-library/no-unnecessary-act
+    await act(async () => {
+      fireEvent.click(addMappingButton);
+    });
+
+    const fieldSecond = await screen.findByLabelText(/select customFieldMappings.1.field/i);
+    const typeSecond = await screen.findByLabelText(/select customFieldMappings.1.type/i);
+    const submitButton = await screen.findByLabelText('Submit');
+
+    // eslint-disable-next-line testing-library/no-unnecessary-act
+    await act(async () => {
+      await selectItem(typeSecond, 'String type');
+      await selectItem(fieldSecond, 'http_method');
+
+      fireEvent.click(submitButton);
+    });
+
+    expect(mockSubmit).not.toHaveBeenCalled();
   });
 });
