+++ graylog2-web-interface/src/pages/IndexSetFieldTypeProfileEditPage.tsx
@@ -15,17 +15,16 @@
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
 import React, { useEffect } from 'react';
-import { Loader } from '@mantine/core';
 import { useNavigate } from 'react-router-dom';
 
-import { DocumentTitle, PageHeader } from 'components/common';
+import { DocumentTitle, PageHeader, Spinner } from 'components/common';
 import DocsHelper from 'util/DocsHelper';
 import { LinkContainer } from 'components/common/router';
 import Routes from 'routing/Routes';
 import { Button, Col, Row } from 'components/bootstrap';
-import useProfile from 'components/indices/IndexSetFiledTypeProfiles/hooks/useProfile';
+import useProfile from 'components/indices/IndexSetFieldTypeProfiles/hooks/useProfile';
 import useParams from 'routing/useParams';
-import EditProfile from 'components/indices/IndexSetFiledTypeProfiles/EditProfile';
+import EditProfile from 'components/indices/IndexSetFieldTypeProfiles/EditProfile';
 import useCurrentUser from 'hooks/useCurrentUser';
 
 const IndexSetFieldTypeProfileEditPage = () => {
@@ -44,29 +43,27 @@ const IndexSetFieldTypeProfileEditPage = () => {
 
   return (
     <DocumentTitle title="Edit Index Set Field Type Profile">
-      <div>
-        <PageHeader title={`Edit  "${data?.name}" Index Set Field Type Profiles`}
-                    documentationLink={{
-                      title: 'Index model documentation',
-                      path: DocsHelper.PAGES.INDEX_MODEL,
-                    }}
-                    topActions={(
-                      <LinkContainer to={Routes.SYSTEM.INDICES.FIELD_TYPE_PROFILES.OVERVIEW}>
-                        <Button bsStyle="info">Profiles</Button>
-                      </LinkContainer>
+      <PageHeader title={`Edit  "${data?.name}" Index Set Field Type Profile`}
+                  documentationLink={{
+                    title: 'Index model documentation',
+                    path: DocsHelper.PAGES.INDEX_MODEL,
+                  }}
+                  topActions={(
+                    <LinkContainer to={Routes.SYSTEM.INDICES.FIELD_TYPE_PROFILES.OVERVIEW}>
+                      <Button bsStyle="info">Profiles</Button>
+                    </LinkContainer>
                     )}>
-          <span>
-            With index set field type profiles you can bundle up custom field types into profiles.
-            Then you can assign this profile to any of index set.
-            {`On this page you can modify "${data?.name}" field type profiles configuration.`}
-          </span>
-        </PageHeader>
-        <Row className="content">
-          <Col md={12}>
-            {isFetched ? <EditProfile profile={data} /> : <Loader />}
-          </Col>
-        </Row>
-      </div>
+        <span>
+          With index set field type profiles you can bundle up custom field types into profiles.
+          Then you can assign this profile to any of index set.
+          On this page you can modify {data?.name} field type profile configuration.
+        </span>
+      </PageHeader>
+      <Row className="content">
+        <Col md={12}>
+          {isFetched ? <EditProfile profile={data} /> : <Spinner />}
+        </Col>
+      </Row>
     </DocumentTitle>
   );
 };
