+++ graylog2-web-interface/src/pages/IndexSetFieldTypeProfilesPage.tsx
@@ -23,8 +23,8 @@ import { Row, Col } from 'components/bootstrap';
 import DocsHelper from 'util/DocsHelper';
 import Routes from 'routing/Routes';
 import useCurrentUser from 'hooks/useCurrentUser';
-import ProfilesList from 'components/indices/IndexSetFiledTypeProfiles/ProfilesList';
-import CreateProfileButton from 'components/indices/IndexSetFiledTypeProfiles/CreateProfileButton';
+import ProfilesList from 'components/indices/IndexSetFieldTypeProfiles/ProfilesList';
+import CreateProfileButton from 'components/indices/IndexSetFieldTypeProfiles/CreateProfileButton';
 import { IndicesPageNavigation } from 'components/indices';
 
 const IndexSetFieldTypeProfilesPage = () => {
@@ -41,26 +41,24 @@ const IndexSetFieldTypeProfilesPage = () => {
   return (
     <DocumentTitle title="Index Set Field Type Profiles">
       <IndicesPageNavigation />
-      <div>
-        <PageHeader title="Configure Index Set Field Type Profiles"
-                    documentationLink={{
-                      title: 'Index model documentation',
-                      path: DocsHelper.PAGES.INDEX_MODEL,
-                    }}
-                    actions={<CreateProfileButton />}>
-          <span>
-            With index set field type profiles you can bundle up custom field types into profiles.
-            Then you can assign this profile to any of index set.
-            You can view and edit existing profiles or create new ones
-          </span>
-        </PageHeader>
+      <PageHeader title="Configure Index Set Field Type Profiles"
+                  documentationLink={{
+                    title: 'Index model documentation',
+                    path: DocsHelper.PAGES.INDEX_MODEL,
+                  }}
+                  actions={<CreateProfileButton />}>
+        <span>
+          With index set field type profiles you can bundle up custom field types into profiles.
+          Then you can assign this profile to any of index set.
+          You can view and edit existing profiles or create new ones
+        </span>
+      </PageHeader>
 
-        <Row className="content">
-          <Col md={12}>
-            <ProfilesList />
-          </Col>
-        </Row>
-      </div>
+      <Row className="content">
+        <Col md={12}>
+          <ProfilesList />
+        </Col>
+      </Row>
     </DocumentTitle>
   );
 };
